refactor: refactor a lot
