+++ src/so_vits_svc_fork/modules/onnx/__init__.py
@@ -1,0 +1,3 @@
+from ._export import onnx_export
+
+__all__ = ["onnx_export"]
